+++ src/main/java/video/api/java/sdk/domain/analytics/Info.java
@@ -7,9 +7,9 @@ public class Info {
     public final String              sessionId;
     public final Calendar            loadedAt;
     public final Calendar            endedAt;
-    public final Map<String, String> metadata;
+    public final Map<String, Object> metadata;
 
-    public Info(String sessionId, Calendar loadedAt, Calendar endedAt, Map<String, String> metadata) {
+    public Info(String sessionId, Calendar loadedAt, Calendar endedAt, Map<String, Object> metadata) {
         this.sessionId = sessionId;
         this.loadedAt  = loadedAt;
         this.endedAt   = endedAt;