+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 
-## [0.6.0] - 2020-10-15
+## [0.6.0] - 2020-10-16
 
 ### Changed
 - Replaced JSONObject.has() with !JSONObject.isNull()